Add docstrings and seperate things. Closes #214
Bit of a big change to be honest. This is something I've been working on for a while really. Adds doc strings to pretty much everything, adds slots to all the classes, separates Connection into `BaseConnection`, `BaseReadSync`, and `BaseWriteSync` (don't worry `Connection` is still there, just subclasses all of these.) so that there can be separate `BaseReadAsync` and `BaseWriteAsync`, so that `TCPAsyncSocketConnection` and `UDPAsyncSocketConnection` can exist and not have tons of duplicated code.

Also adds unused `read_varlong` and `write_varlong` functions, this could be removed but why not.

This might make python 3.7 not work right because I got rid of some odd stuff with the `BytesConvertable` type, but I haven't tested.