Revert "Run black as a check, not as an autoformatter"
This reverts commit 0d996de8b96fd8be7c1db9a86b84a30245419f14.

Turns out precommit is smart enough to detect file edits and when it
does, it causes the check to fail, even though it itself didn't end up
in a non-zero exit code. This means we can actually run black here as a
full auto-formatter, rather than purely as a check.