Add a temporary fix for tests on mac-os For some reason, when running tests on macos, DNS A record resulution is failing with NXDOMAIN when trying to resolve for 'localhost', instead of properly returning '127.0.0.1'. This issue wasn't previously catched, before the DNS resolving implementation was catching this NXDOMAIN as a way to handle for host already being an IP.
This is a temporary fix which adds try-except for NXDOMAIN and sets the
IP to host (just like in the previous implementation), which bypasses
this failure. But this should be removed as soon as we find the root
cause of this mac-os test failure (I don't suppose mac-os actually can't
resolve 'localhost' to IP, though it would seem that way from the
workflow).