Small improvements to excluded lines in coverage
Why space added? Actual comment is with space,
[source](https://coverage.readthedocs.io/en/6.4.1/excluding.html).

Why three dots? This will exclude overloads as it is used in #306, and
manually uncovered by `no cover` comments.

~~Why was removed `:` in `if TYPE_CHECKING`? This if block can be also
`if TYPE_CHECKING is True` or `if TYPE_CHECKING is False`, so this will
escape all of usages like, what we actually expect.~~ Change was reverted,
[see details](https://github.com/py-mine/mcstatus/pull/357#discussion_r922888879).

---

Another changes was proposed by @ItsDrike
[here](https://github.com/py-mine/mcstatus/pull/357#discussion_r922888879).