Use typing.Any for obj type to deprecated
Because of a compatibility breaking update from pyright type-checker,
the types used in the concrete implementation of the deprecated
decorator were marked as incompatible with the signature of the
typing overloads.

Since the actual type of `obj` would be pretty complex if we wanted to
actually make it compatible with all of the various overload usages it
supports, instead of creating a compatible fully typed signature for the
concrete implementation, this simply changes the type-annotation of this
argument to `typing.Any`.

This is not a big issue for internal consistency, since the `obj`
parameter will immediately get passed into an internal `decorate`
function, which is fully annotated and contains the actual logic of the
function, so this type signature in the concrete version isn't actually
that important for us internally, and externally (when using the
function), we will only see the possible overloads, which are still
properly type-hinted.

This greatly reduces the type complexity that would be required to make
the concrete version compatible, and also prevents potential future
issues with more breaking changes in the type-checker.