Add flake8 tools to enforce these changes
In PEP-585, commonly used abcs in the standard library and built-in
types were made generic (such as the collections.abc classes, or the
built-in list, type, dict, ... classes).

Since this PEP only applies to python 3.9+, we can't use these generics
on runtime, because our project is 3.7+, however with
__future__.annotations compiler setting, runtime errors won't be
produced, and type-checkers can already understand this new syntax.

To enforce these changes, flake8-585 (for 3.9+) was added as a dev
dependency, and to ensure that when using these changes, we stay 3.7
compatible, this also adds flake8-future-annotations, which will produce
a flake8 error when trying to use these generics without the
__future__.annotations import.