fix command-line call order for bungeecord servers.
according to @py-mine-bot's research ( https://github.com/py-mine/mcstatus/issues/174#issuecomment-1041004712 )
a client is first expected to request status and then ping.
most proxy's ignore this and allow a ping anytime, but bungeecord servers will not respond with a ping if a status was not asked for first.

this causes a result like:

```
# mcstatus.py play.mineclub.com json
{"online": false}
```

after swapping the order of the status/ping we get:

```
# mcstatus.py play.mineclub.com json
{"online": true, "version": "1.19.1", "protocol": 47, "motd": "\u00a7b\u00a7l            Mineclub \u00a77- \u00a7d\u00a7lBeta 0.6.4\u00a7r\n\u00a76             Summer Unobtainable Event!", "player_count": 163, "player_max": 164, "players": [], "ping": 31.002}
```

note, the actual call to ping() is not even needed as the ping is also done with a .status() call so we can get the ping result from `status_res.latency`

this should help with issues like #174, #92 and #112 so that the program works as expected when called from the command line when ran aginst a bungecoorded server.