Fix poetry 1.2 support
Release of poetry 1.2, introduced a regression which broke pip --user
installs. These types of installs were the main way we performed
instalations in the CI workflows, as they made it more convenient to
control the location, availability and caching of packages.

Poetry's team does not recognize this as a supported use case, so major
changes were required to get everything working again. Most of the
changes were consolidated into HassanAbouelela/setup-python which we now
use to set up python and poetry for us.