Move to using poetry dependency groups
This moves the current monolithic dependency structure which only
distinguishes between dev dependencies and main project dependencies to
a more modular structure where the dev dependencies aren't all clumped
together.

By default, these dependency groups will always get installed with
`poetry install` unless it's an optional dependency group, in which case
it needs to be included manually using `poetry install --with
my-optional-group`. Similarly, we can also exclude a dependency group
using `poetry install --without my-group` if we don't need it. For
example we don't need to install the dependencies for linting in our
testing workflow, so we can simply use `poetry install --without lint`
to exclude this group.

Note that these groups are internal to poetry only, and unlike extras,
they won't affect releases, which will still only include the
dependencies from main group. This change therefore shouldn't have any
effect on the end-users.

Because we can have these optional groups, which aren't installed by
default, it finally allows us to specify a locked version of some
dependencies we need in workflows (instead of just running `pip install`
in the workflow) and have poetry manage it, without just adding it to
the dev dependencies and have it needlessly install for everyone (like
with `tox-gh-actions`). Other than just with workflows, it also allows
us to separate release dev dependencies (like `twine`) as most people
working on the library won't actually be making releases.

Dependency groups are a relatively new feature in poetry, and were
introduced as a part of the 1.2 release. For more info about them, see
the documentation:
- https://python-poetry.org/docs/managing-dependencies/#dependency-groups
- https://python-poetry.org/docs/pyproject/#dependencies-and-dependency-groups
- https://python-poetry.org/docs/cli/#install