Remove tox and use poe for task execution instead
Tox served us well, but it is something we have discussed replacing. The
ecosystem for version selection is better than it was; poetry now has
the ability to easily switch versions with `poetry env`. The format of
`tox.ini` is not as easy as poe's configuration in `pyproject.toml`;
this should simplify adding and maintaining tasks. Additionally, tox
would occasionally release breaking changes such as hiding environment
variables; switching to a simpler task execution command should help
minimize the amount of breaking changes because it is a smaller
codebase.