Add tests for our CLI (#877)
* tests: add basic end-to-end tests for CLI

currently makes use of demo.mcstatus.io as a test server,
but it would be good to run a local instance maybe.

also adds output tests for --help, allowing this to be used
in documentation.

* fix badly formatted string

* normalise --help output

* Make tests fully offline and improve coverage

* Address review

* Fix tests

* Use repr for common errors

* Fix CI

* Raise `TimeoutError` instead of `socket.timeout` in mock

In 3.12 `socket.timeout` is an alias to `TimeoutError`, but in previous
versions they are different.

* Catch `TimeoutError` in CLI as one of the common exceptions

* Clarify comment as requested in review

---------

Co-authored-by: rina <k@rina.fyi>